requirements.txt
locust>=2.0.0
python-dotenv>=0.19.0
har2locust>=0.6.0  # For HAR file automation 